moved integration test out of the unit module
